+++ ta4j-core/src/main/java/org/ta4j/core/BaseStrategy.java
@@ -32,7 +32,7 @@ import org.slf4j.LoggerFactory;
 public class BaseStrategy implements Strategy {
 
     /** The logger. */
-    protected final transient Logger log = LoggerFactory.getLogger(getClass());
+    protected final Logger log = LoggerFactory.getLogger(getClass());
 
     /** The class name. */
     private final String className = getClass().getSimpleName();