+++ ta4j-core/src/main/java/org/ta4j/core/criteria/ReturnOverMaxDrawdownCriterion.java
@@ -32,9 +32,12 @@ import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 
 /**
- * Reward risk ratio criterion, defined as the {@link ReturnCriterion gross
- * return} over the {@link MaximumDrawdownCriterion maximum drawdown}, returned
- * in decimal format.
+ * Reward risk ratio criterion (also known as "RoMaD"), returned in decimal
+ * format.
+ * 
+ * <pre>
+ * RoMaD = {@link ReturnCriterion gross return (with base)} / {@link MaximumDrawdownCriterion maximum drawdown}
+ * </pre>
  */
 public class ReturnOverMaxDrawdownCriterion extends AbstractAnalysisCriterion {