+++ ta4j-core/src/main/java/org/ta4j/core/indicators/MACDIndicator.java
@@ -24,6 +24,7 @@
 package org.ta4j.core.indicators;
 
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.numeric.NumericIndicator;
 import org.ta4j.core.num.Num;
 
 /**
@@ -81,6 +82,22 @@ public class MACDIndicator extends CachedIndicator<Num> {
      */
     public EMAIndicator getLongTermEma() {
         return longTermEma;
+    }
+
+    /**
+     * @param barCount of signal line
+     * @return signal line for this MACD indicator
+     */
+    public EMAIndicator getSignalLine(int barCount) {
+        return new EMAIndicator(this, barCount);
+    }
+
+    /**
+     * @param barCount of signal line
+     * @return histogram of this MACD indicator
+     */
+    public NumericIndicator getHistogram(int barCount) {
+        return NumericIndicator.of(this).minus(getSignalLine(barCount));
     }
 
     @Override