+++ ta4j-core/src/main/java/org/ta4j/core/indicators/SMAIndicator.java
@@ -24,6 +24,7 @@
 package org.ta4j.core.indicators;
 
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.helpers.RunningTotalIndicator;
 import org.ta4j.core.num.Num;
 
 /**
@@ -34,8 +35,8 @@ import org.ta4j.core.num.Num;
  */
 public class SMAIndicator extends CachedIndicator<Num> {
 
-    private final Indicator<Num> indicator;
     private final int barCount;
+    private RunningTotalIndicator previousSum;
 
     /**
      * Constructor.
@@ -45,21 +46,22 @@ public class SMAIndicator extends CachedIndicator<Num> {
      */
     public SMAIndicator(Indicator<Num> indicator, int barCount) {
         super(indicator);
-        this.indicator = indicator;
+        this.previousSum = new RunningTotalIndicator(indicator, barCount);
         this.barCount = barCount;
     }
 
     @Override
     protected Num calculate(int index) {
-        Num sum = zero();
-        for (int i = Math.max(0, index - barCount + 1); i <= index; i++) {
-            sum = sum.plus(indicator.getValue(i));
-        }
-
         final int realBarCount = Math.min(barCount, index + 1);
+        final var sum = partialSum(index);
         return sum.dividedBy(numOf(realBarCount));
     }
 
+    private Num partialSum(int index) {
+        return this.previousSum.getValue(index);
+    }
+
+    /** @return {@link #barCount} */
     @Override
     public int getUnstableBars() {
         return barCount;