+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/ClosePriceDifferenceIndicator.java
@@ -28,8 +28,11 @@ import org.ta4j.core.indicators.CachedIndicator;
 import org.ta4j.core.num.Num;
 
 /**
- * Indicator that calculates the difference between the close prices of the
- * current bar and the previous bar.
+ * Calculates the difference between the current and the previous close price.
+ * 
+ * <pre>
+ * ClosePriceDifference = currentBarClosePrice - previousBarClosePrice
+ * </pre>
  */
 public class ClosePriceDifferenceIndicator extends CachedIndicator<Num> {
 
@@ -61,13 +64,9 @@ public class ClosePriceDifferenceIndicator extends CachedIndicator<Num> {
         return currentBarClosePrice.minus(previousBarClosePrice);
     }
 
-    /**
-     * {@inheritDoc}
-     * <p>
-     * This indicator is always stable, so it returns 0.
-     */
+    /** @return {@code 1} */
     @Override
     public int getUnstableBars() {
-        return 0;
+        return 1;
     }
 }