+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/SumIndicator.java
@@ -28,31 +28,33 @@ import org.ta4j.core.indicators.CachedIndicator;
 import org.ta4j.core.num.Num;
 
 /**
- * Sum indicator.
- *
- * I.e.: operand0 + operand1 + ... + operandN
+ * Calculates the sum of all indicator values.
+ * 
+ * <pre>
+ * Sum = summand0 + summand1 + ... + summandN
+ * </pre>
  */
 public class SumIndicator extends CachedIndicator<Num> {
 
-    private final Indicator<Num>[] operands;
+    private final Indicator<Num>[] summands;
 
     /**
-     * Constructor. (operand0 plus operand1 plus ... plus operandN)
+     * Constructor.
      *
-     * @param operands the operand indicators for the sum
+     * @param summands the indicators ​​to be summed
      */
     @SafeVarargs
-    public SumIndicator(Indicator<Num>... operands) {
+    public SumIndicator(Indicator<Num>... summands) {
         // TODO: check if first series is equal to the other ones
-        super(operands[0]);
-        this.operands = operands;
+        super(summands[0]);
+        this.summands = summands;
     }
 
     @Override
     protected Num calculate(int index) {
-        Num sum = numOf(0);
-        for (Indicator<Num> operand : operands) {
-            sum = sum.plus(operand.getValue(index));
+        Num sum = zero();
+        for (Indicator<Num> summand : summands) {
+            sum = sum.plus(summand.getValue(index));
         }
         return sum;
     }