+++ ta4j-core/src/test/java/org/ta4j/core/indicators/supertrend/SuperTrendIndicatorTest.java
@@ -67,6 +67,12 @@ public class SuperTrendIndicatorTest extends AbstractIndicatorTest<Indicator<Num
         bars.add(new MockBar(19.00, 18.35, 19.41, 18.01, numFunction));
         bars.add(new MockBar(19.89, 6.36, 20.22, 6.21, numFunction));
         bars.add(new MockBar(19.28, 10.34, 20.58, 10.11, numFunction));
+        // this mock bar exemplify an edge case, the close price is the same as the
+        // previous Super Trend value
+        bars.add(new MockBar(19.28, 22.78938583966133, 23.58, 10.11, numFunction));
+        bars.add(new MockBar(19.28, 10.34, 20.58, 10.11, numFunction));
+        bars.add(new MockBar(10.34, 9.83, 12.80, 8.83, numFunction));
+        bars.add(new MockBar(11.83, 7.35, 11.41, 5.01, numFunction));
 
         data = new MockBarSeries(bars);
     }
@@ -78,6 +84,17 @@ public class SuperTrendIndicatorTest extends AbstractIndicatorTest<Indicator<Num
         assertNumEquals(this.numOf(15.730621000000003), superTrendIndicator.getValue(4));
         assertNumEquals(this.numOf(17.602360938100002), superTrendIndicator.getValue(9));
         assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(14));
+    }
 
+    @Test
+    public void regressionTestOnZeroSuperTrendValueWhenClosePriceIsEqualToPreviousSuperTrendValue() {
+        // bug: https://github.com/ta4j/ta4j/issues/1120
+        SuperTrendIndicator superTrendIndicator = new SuperTrendIndicator(data);
+
+        assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(14));
+        assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(15));
+        assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(16));
+        assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(17));
+        assertNumEquals(this.numOf(22.78938583966133), superTrendIndicator.getValue(18));
     }
 }